Expose artworks_count and artists_count in myCollectionInfo (#4074)
* feat: Expose artwork and artist count in myCollectionInfo

* bring back my collection fields

* update schema
