+++ src/schema/v2/me/index.ts
@@ -46,8 +46,9 @@ import {
 import LotStanding from "./lot_standing"
 import LotStandings from "./lot_standings"
 import { MyBids } from "./myBids"
-import { MyCollection, MyCollectionInfo } from "./myCollection"
+import { MyCollection } from "./myCollection"
 import MyCollectionAuctionResults from "./myCollectionAuctionResults"
+import { MyCollectionInfo } from "./myCollectionInfo"
 import { myLocationType } from "./myLocation"
 import { NewWorksByInterestingArtists } from "./newWorksByInterestingArtists"
 import { ManagedPartners } from "./partners"
